Code saved locally
this is the code we actually working on in vscode. Victor is new, drivetrain is updated to traditional tank drive instead of drone drive, constants was majorly updated, and shooter as well. OI and RobotMap are gone, commands is gone with colorfinder, examplecommand, navxseeker, rotationcontrol, and seekvisiontarget. The subsystems CPI, colorsensor, examplesubsystem, navx, encoder, limelight, and pneumatics have all been dealt away with. Oh, and some things have been changed from public to private to lessen the clutter.